Refactors bot
