+++ src/domain/anime-detector/module.ts
