+++ src/domain/announcer/calendar/service.ts
@@ -1,15 +1,19 @@
 import fetch from 'node-fetch';
-import { LoggingService } from '../utils/logging';
-import { AcademicWeeksParser, AcademicCalendarService, AcademicDeadlinesParser, AcademicCalendar, Deadline } from './types';
-import { getWeekIndex } from './week';
+import { LoggingService } from '../../../utils/logging';
+import { academicDeadlinesParser } from './deadlines/parser';
+import { AcademicCalendarService, AcademicCalendar, Deadline } from './types';
+import { academicWeeksParser, getWeekIndex } from './weeks/parser';
 
 const TEACHING_WEEKS_URL = 'https://ipoint.uwa.edu.au/app/answers/detail/a_id/1405/~/2016-dates-and-teaching-weeks';
 const CS_MARKS_URL = 'https://secure.csse.uwa.edu.au/run/cssubmit';
 
 const zip = <T1, T2>(arr1: T1[], arr2: T2[]): [T1, T2][] => arr1.map((_, idx) => [arr1[idx], arr2[idx]]);
 
-export const academicCalendarService = ({ log }: LoggingService, weeksParser: AcademicWeeksParser, deadlinesParser: AcademicDeadlinesParser, filterPostgrad = true): AcademicCalendarService => ({
-  fetchCalendar: async () => {
+export const academicCalendarService = ({ log }: LoggingService, filterPostgrad = true): AcademicCalendarService => {
+  const weeksParser = academicWeeksParser();
+  const deadlinesParser = academicDeadlinesParser();
+
+  return async () => {
     const fetchWeeks = async () => {
       const weeksResult = await fetch(TEACHING_WEEKS_URL);
       const weeksHtml = await weeksResult.text();
@@ -31,9 +35,9 @@ export const academicCalendarService = ({ log }: LoggingService, weeksParser: Ac
     const deadlines = await fetchDeadlines();
 
     const calendar: AcademicCalendar = { weeks };
-    deadlines.forEach(deadline => calendar.weeks[getWeekIndex(deadline.date)].deadlines.push(deadline));
+    deadlines.forEach(deadline => calendar.weeks[getWeekIndex(deadline.date)]?.deadlines.push(deadline));
 
     log('info', 'Fetched and parsed an academic calendar', { title: 'Academic Calendar Service', data: { calendar } });
     return calendar;
-  },
-});
+  };
+};
