+++ src/domain/announcer/module.spec.ts
@@ -1,61 +1,59 @@
-import { AcademicCalendarService } from '../../academic-calendar/types';
+import { AcademicCalendarService } from './calendar/types';
 import { mockLogger } from '../../utils/logging';
 import { BotActionType } from '../action-types';
-import { AnnouncerConfig, announcerModule } from './announcer';
+import { AnnouncerConfig, announcerModule } from './module';
 
-const mockCalendarService = (): AcademicCalendarService => ({
-  fetchCalendar: async () => ({
-    weeks: {
-      ['2020-02-24']: {
-        type: 'teaching',
-        week: 1,
-        semester: 1,
-        date: new Date('2020-02-24T00:00:00.000Z'),
-        deadlines: [],
-      },
-      ['2020-03-02']: {
-        type: 'teaching',
-        week: 2,
-        semester: 1,
-        date: new Date('2020-03-02T00:00:00.000Z'),
-        deadlines: [{ unit: 'TEST1001', title: 'Test 1', date: new Date('2020-03-03T23:59:00.000+08:00') }],
-      },
-      ['2020-03-09']: {
-        type: 'study-break',
-        date: new Date('2020-03-09T00:00:00.000Z'),
-        deadlines: [],
-      },
-      ['2020-03-16']: {
-        type: 'exam',
-        date: new Date('2020-03-16T00:00:00..000Z'),
-        deadlines: [],
-      },
-      ['2020-07-27']: {
-        type: 'teaching',
-        week: 1,
-        semester: 2,
-        date: new Date('2020-07-27T00:00:00.000Z'),
-        deadlines: [],
-      },
-      ['2020-08-03']: {
-        type: 'teaching',
-        date: new Date('2020-08-03T00:00:00.000Z'),
-        week: 2,
-        semester: 2,
-        deadlines: [{ unit: 'TEST2002', title: 'Test 2', date: new Date('2020-08-05T12:54:00.000+08:00') }],
-      },
-      ['2020-08-10']: {
-        type: 'study-break',
-        date: new Date('2020-08-10T00:00:00.000Z'),
-        deadlines: [],
-      },
-      ['2020-08-17']: {
-        type: 'exam',
-        date: new Date('2020-08-17T00:00:00.000Z'),
-        deadlines: [],
-      },
+const mockCalendarService = (): AcademicCalendarService => async () => ({
+  weeks: {
+    ['2020-02-24']: {
+      type: 'teaching',
+      week: 1,
+      semester: 1,
+      date: new Date('2020-02-24T00:00:00.000Z'),
+      deadlines: [],
     },
-  }),
+    ['2020-03-02']: {
+      type: 'teaching',
+      week: 2,
+      semester: 1,
+      date: new Date('2020-03-02T00:00:00.000Z'),
+      deadlines: [{ unit: 'TEST1001', title: 'Test 1', date: new Date('2020-03-03T23:59:00.000+08:00') }],
+    },
+    ['2020-03-09']: {
+      type: 'study-break',
+      date: new Date('2020-03-09T00:00:00.000Z'),
+      deadlines: [],
+    },
+    ['2020-03-16']: {
+      type: 'exam',
+      date: new Date('2020-03-16T00:00:00..000Z'),
+      deadlines: [],
+    },
+    ['2020-07-27']: {
+      type: 'teaching',
+      week: 1,
+      semester: 2,
+      date: new Date('2020-07-27T00:00:00.000Z'),
+      deadlines: [],
+    },
+    ['2020-08-03']: {
+      type: 'teaching',
+      date: new Date('2020-08-03T00:00:00.000Z'),
+      week: 2,
+      semester: 2,
+      deadlines: [{ unit: 'TEST2002', title: 'Test 2', date: new Date('2020-08-05T12:54:00.000+08:00') }],
+    },
+    ['2020-08-10']: {
+      type: 'study-break',
+      date: new Date('2020-08-10T00:00:00.000Z'),
+      deadlines: [],
+    },
+    ['2020-08-17']: {
+      type: 'exam',
+      date: new Date('2020-08-17T00:00:00.000Z'),
+      deadlines: [],
+    },
+  },
 });
 
 describe('announcer-module', () => {
