+++ src/domain/cowsay/module.ts
